Add files via upload
The subject of our project is a healthcare monitoring and consulting system. Admin, customer, and consultant are the three components. To access their panels, they must first log in. We chose a web-based solution since it is easy to use. The group's consensus is that everyone involved in the initiative is benefitting from the web portal. This will help in the public and private sectors, which is where the main issue lies.